refactor(transactions): extract TransactionForm shared component
- Extract form UI from add.tsx into reusable TransactionForm component
- TransactionForm accepts all state as props (controlled component)
- Add optional onDelete prop for delete button support
- Refactor add.tsx to thin wrapper connecting Zustand store to TransactionForm
- Export TransactionForm from features/transactions barrel
